Fix error 500 on performace test

+++ SafeguardSpring/src/main/java/com/dqtri/mango/safeguard/controller/SubmissionController.java
@@ -207,6 +207,9 @@ public class SubmissionController {
     @PreAuthorize("hasAnyRole('ADMIN', 'SUBMITTER') and hasPermission(#id, @submissionOwner)")
     public ResponseEntity<Void> deleteSubmission(@PathVariable("id") Long id) {
         Submission submission = getSubmissionOrElseThrow(id);
+        if (!Status.DRAFT.equals(submission.getStatus())) {
+            throw new LockedException("The submission is locked");
+        }
         submissionRepository.delete(submission);
         return ResponseEntity.noContent().build();
     }
